requirements.txt
psycopg2-binary==2.8.5
django-crispy-forms==1.9.0
pytz==2020.1